refactor: Collection to CollectionPage
